fix: should not alter votable proposals before verification
Reintroducing the logic that was removed on ac29227bab4afbbf35391c391f9d5c2f82a33a51 and also removing the `pop`. We do this because popping and then not managing to perform the transaction would result in the proposal getting deleted and then not trying to vote for it again.